requirements.txt
numpy
pandas
scipy
tqdm
datasets
huggingface-hub
sentence-transformers>=4.1.0
torch>=1.11.0
transformers>=4.41.0
scikit-learn
matplotlib>=3.5.0
seaborn>=0.12.0
joblib>=1.2.0
binarytree>=6.5.1
psutil>=5.9.0
einops>=0.8.1
pyarrow>=10.0.0
PyYAML>=6.0 
pydantic>=2.0
numba>=0.60.0
